+++ test/faker/cannabis_test.exs
@@ -1,0 +1,6 @@
+defmodule Faker.CannabisTest do
+  use ExUnit.Case, async: true
+
+  doctest Faker.Cannabis
+  doctest Faker.Cannabis.En
+end