+++ test/faker/finance_test.exs
@@ -1,0 +1,5 @@
+defmodule Faker.FinanceTest do
+  use ExUnit.Case, async: true
+
+  doctest Faker.Finance.Stock
+end